requirements.txt
mysql-connector-python==8.0.5
python-telegram-bot==12.8